// src/app/shared/models/permissions-checking.model.ts
export class PermissionsCheckingModel {
    isAddInvestigation: boolean;
    isViewInvestigation: boolean;
    isUpdateInvestigation: boolean;
    isCloseInvestigation: boolean;
    isNewRequest: boolean;
    isViewDocumentRequests: boolean;
    isUploadDocument: boolean;
    isViewSentDocument: boolean;
    isViewReceivedDocument: boolean;
    isReminderRequest: boolean;
    isCloseRequest: boolean;
    isAddNotes: boolean;
    isEditNotes: boolean;
    isViewNotes: boolean;
    isDeleteNotes: boolean;
    isAttachClaims: boolean;
    isUpdatePendCodeClaim: boolean;
    isReOpenInvestigation: boolean;
    isReOpenClaim: boolean;
    isNewMiscDocuments: boolean;
    isDownloadMiscDocument: boolean;
    isAttachMiscDocument: boolean;
    isDeleteMiscDocument: boolean;
    isViewClaims: boolean;
    isViewMiscDocuments: boolean;
    isAddPermissions: boolean;
}